feat(spbj): add blackjack-core feature which only returns the data that should be known to the player
ISinglePlayerGameManager has new method: getPublicGameManager(), which essentially returns the same gamestate as before with the addition that if dealer's hand is hidden it then removes the hidden cards from the returned object. Meaning the frontend will not receive the information of what cards the dealer has (except the upcard) until the dealer has to show their hand.